refactor(#3038): VoiceBargeInRuntime S4 — extract routing resolution cluster
S4 scope: extract the next cohesive VoiceBargeInRuntime subcluster from the S1 concern map without behavior changes. After S3, routing/dispatch is the largest remaining aggregate, but moving it wholesale would exceed the child threshold and mix dispatch, handoff, processing, deferred drain, and route resolution. This slice picks the largest clean routing/dispatch ownership boundary: routing resolution and reverse lookup.

Cluster choice: live-cut playback and the TTS pipeline remain smaller cohesive candidates from the map, while this routing-resolution cluster moves cleanly and is larger. `cached_config` stays in the root because the S1 map classifies it as lifecycle/config and playback children also use it.

Cluster boundary: moved `resolve_effective_foreground_config`, `resolve_voice_background_channel_for_source`, `active_barge_in_mailbox_channel`, `voice_channel_for_background`, `resolve_voice_turn_target`, `bind_routed_voice_context`, `voice_turn_guild_id`, and `ask_for_agent` verbatim into `src/services/discord/voice_barge_in/routing.rs` as an inherent impl child module. The only moved-method signature edits are visibility-only: parent/sibling callers require `pub(super)` on `resolve_effective_foreground_config`, `resolve_voice_background_channel_for_source`, `active_barge_in_mailbox_channel`, `resolve_voice_turn_target`, `voice_turn_guild_id`, and `ask_for_agent`. A root `settings` import keeps the moved method body path `super::settings::resolve_role_binding` unchanged at the child boundary.

No await point, lock scope, spawn, map-entry, cancellation, channel-ordering, or queue-mutation behavior changed. No struct fields were widened. Existing cfg behavior is unchanged.

LoC accounting: `voice_barge_in.rs` production LoC drops 3710 -> 3332 (-378). New child `voice_barge_in/routing.rs` is 383 production LoC, below the 1000-line giant threshold, so only the root ratchet is lowered.

Token proof: normalized method-token comparison against HEAD is identical after visibility-only normalization. Per moved method:

| Method | Method tokens | Body tokens | Normalized SHA |
| --- | ---: | ---: | --- |
| `resolve_effective_foreground_config` | 343 -> 343 | 323 -> 323 | `a105f338bb4a` |
| `resolve_voice_background_channel_for_source` | 80 -> 80 | 61 -> 61 | `e4d0b4ad5573` |
| `active_barge_in_mailbox_channel` | 107 -> 107 | 80 -> 80 | `db5b0d6c64ae` |
| `voice_channel_for_background` | 663 -> 663 | 636 -> 636 | `09c5b6029fec` |
| `resolve_voice_turn_target` | 694 -> 694 | 665 -> 665 | `c68d6e485cbe` |
| `bind_routed_voice_context` | 70 -> 70 | 53 -> 53 | `a38368b7d92c` |
| `voice_turn_guild_id` | 81 -> 81 | 59 -> 59 | `ddd7231f5921` |
| `ask_for_agent` | 99 -> 99 | 78 -> 78 | `888e2e43a7fd` |
| **Total** | **2137 -> 2137** | **1955 -> 1955** | |

Remaining-map status after S4:

- routing/dispatch still owns the heavier dispatch path in the root: channel text reply, processing transcript, transcript announcement, background handoff, start voice turn, completed utterance processing, and deferred drain.
- playback remaining in the root is the live-cut subcluster: playback reset/clear ownership, live PCM observation, monitor lookup, sensitivity mirror, and monitor sensitivity updates.
- TTS pipeline remains in the root: runtime TTS voice rebuild plus progress/ack synthesis.
- lifecycle/config and misc methods remain below the larger cohesive candidates.

Gates run before commit:

- cargo fmt --check
- cargo check --lib
- cargo clippy --lib -- -D warnings
- set -o pipefail; cargo test --lib voice 2>&1 | grep "test result" && cargo test --lib discord 2>&1 | grep "test result"
- python3 scripts/audit_maintainability.py --check
