+++ ARCHITECTURE.md
@@ -525,6 +525,7 @@ src/
 │   │   ├── voice_barge_in/
 │   │   │   ├── final_result_playback.rs
 │   │   │   ├── progress_playback.rs
+│   │   │   ├── routing.rs
 │   │   │   └── stt.rs
 │   │   ├── watchers/
 │   │   │   ├── lifecycle.rs
